refactor: purge auto keyword from src/, use explicit types
The library is deliberately C-leaning and explicit; `auto` is not used in
src/. Convert the remaining 9 sites (all lambdas assigned to auto):

- AES key schedule: the captureless SubWord/RotWord lambdas in ssh_aesgcm,
  ssh_aes256ctr and quic_aead become file-static aes_sub_word/aes_rot_word
  free functions with explicit uint32_t return.
- failsafe: the capturing put/put_u32 JSON appenders become fs_put/fs_put_u32
  statics taking (out, cap, *n) explicitly, matching the statsd app() idiom.
- snmp_ber: the capturing OID emit lambda becomes oid_emit_arc(tmp, cap, *t,
  v, e).

Also const-qualifies aes256_ecb's ctx (S995). Behavior identical; AES KATs,
failsafe and snmp tests pass.

+++ src/network_drivers/presentation/dtls/dtls_handshake.cpp
@@ -96,7 +96,7 @@ size_t dtls_hs_header_parse(const uint8_t *p, size_t len, DtlsHsHeader *out)
     out->frag_length = ((uint32_t)p[9] << 16) | ((uint32_t)p[10] << 8) | p[11];
     if (out->frag_offset + out->frag_length > out->length)
         return 0; // fragment falls outside the declared message
-    if ((size_t)DTLS_HS_HDR_LEN + out->frag_length > len)
+    if (DTLS_HS_HDR_LEN + out->frag_length > len)
         return 0; // fragment bytes truncated
     out->fragment = p + DTLS_HS_HDR_LEN;
     return DTLS_HS_HDR_LEN + out->frag_length;
